requirements.txt
pytest
requests
pydantic
allure-pytest
faker
selenium
webdriver-manager